Editando...

+++ backend/order_service/src/main/java/ucv/app_inventory/order_service/application/dto/ProductSupplierDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class ProductSupplierDTO {
-    private int id;
+    private Long id;
     private Long productId;
     private Long supplierId;
-    private Double price;  // El precio se pasa en la solicitud
+    private Double price;
 }
